+++ functions/_adapters/gtfs/services/departures/types.ts
@@ -1,1 +1,1 @@
-export type GtfsDepartureTuple = [string, string, string, number]; // [trip_id, route_id, headsign, timestamp_ms]
+export type GtfsDepartureTuple = [string, string, string, number] | [string, string, string, number, number]; // [trip_id, route_id, headsign, timestamp_ms, wheelchair_accessible?]